requirements.txt
flask==1.1.2
flask-sqlalchemy==2.4.0
psycopg2==2.8.6
python-logstash-async==2.2.0
flask_cors